revamp delete dialog box

+++ lib/widgets/post_card.dart
@@ -307,30 +307,32 @@ class _PostCardState extends State<PostCard> {
                 onPressed: () {
                   showDialog(
                     context: context,
-                    builder: (context) => Dialog(
-                      child: ListView(
-                        padding: EdgeInsets.symmetric(vertical: 16),
-                        shrinkWrap: true,
-                        children: ['Delete']
-                            .map(
-                              (e) => InkWell(
-                                onTap: () async {
-                                  Navigator.of(context).pop();
-                                  FirestoreMethods().deletePost(
-                                    widget.snap['postId'],
-                                  );
-                                },
-                                child: Container(
-                                  padding: EdgeInsets.symmetric(
-                                    vertical: 12,
-                                    horizontal: 16,
-                                  ),
-                                  child: Text(e),
-                                ),
-                              ),
-                            )
-                            .toList(),
+                    builder: (context) => AlertDialog(
+                      shape: RoundedRectangleBorder(
+                        borderRadius: BorderRadius.circular(16.0),
                       ),
+                      title: Text('Delete Post'),
+                      content: Text(
+                        'Are you sure you want to delete this post?',
+                      ),
+                      actions: [
+                        TextButton(
+                          onPressed: () => Navigator.of(context).pop(),
+                          child: Text('Cancel'),
+                        ),
+                        TextButton(
+                          onPressed: () async {
+                            Navigator.of(context).pop();
+                            await FirestoreMethods().deletePost(
+                              widget.snap['postId'],
+                            );
+                          },
+                          child: Text(
+                            'Delete',
+                            style: TextStyle(color: Colors.red),
+                          ),
+                        ),
+                      ],
                     ),
                   );
                 },
